refactor(assets): src/assets 디렉토리 refactor 및 컴포넌트 내 사용 경로 변경
