+++ src/icons/BreadscrumbArrow.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import forwardArrow from '@assets/icons/svgs/arrows/forward-arrow_triangle.svg?react';
+import forwardArrow from '@assets/svgs/arrows/forward-arrow_triangle.svg?react';
 
 export const BreadscrumbArrow = styled(forwardArrow)`
   width: 14px;
