+++ src/pages/Exhibition/Graduation/GraduationExhibition.tsx
@@ -3,7 +3,7 @@ import { motion } from 'framer-motion';
 
 import Exhibition from '@components/Exhibition/Exhibition';
 
-import graduationBanner from '@assets/images/banners/graduation-banner.jpg';
+import graduationBanner from '@assets/images/graduation-banner.jpg';
 
 import * as S from './GraduationExhibition.styled';
